Make clang-tidy clean and enforce it with pinned clang-tidy 19
The previous lint output was broken: clang-tidy 18 on CI could not parse
C++23 <expected> (errors on every file) and the over-broad check set
(cppcoreguidelines-*, pedantic readability) produced huge noise.

- Curate .clang-tidy to a focused, low-noise set; allow explicit (void) casts
  for bugprone-unused-return-value; WarningsAsErrors so findings fail.
- Fix genuine findings: convert() parameter-name mismatch, size_t->int64_t
  narrowing in build_batch, return-with-void-value in the raw writer, and a
  C-style array in a test.
- CI installs clang-tidy 19 (matches local) so it parses C++23, and the
  clang-tidy step is now enforced (no continue-on-error).

Tree is clang-tidy clean (0 findings); 83 tests pass.

+++ src/adapters/batch.cpp
@@ -289,7 +289,8 @@ std::expected<RecordBatch, std::string> build_batch(BatchBuilder& builder) {
         fields.push_back(arrow::field(column.schema.name, arrow_type(column.schema.type)));
         reset(column);
     }
-    auto data = arrow::RecordBatch::Make(arrow::schema(fields), builder.n_rows, arrays);
+    auto data = arrow::RecordBatch::Make(arrow::schema(fields),
+                                         static_cast<std::int64_t>(builder.n_rows), arrays);
     return RecordBatch{data, std::exchange(builder.n_rows, 0)};
 }
